+++ src/overlaybd/gzindex/test/test.cpp
@@ -139,7 +139,7 @@ private:
         if (gzdata == nullptr) {
             LOG_ERRNO_RETURN(0, -1, "failed to create `", fn_gzdata);
         }
-        if (gzdata->pwrite(gzbuf, gzlen, 0) != gzlen) {
+        if (gzdata->pwrite(gzbuf, gzlen, 0) != (ssize_t)gzlen) {
             LOG_ERRNO_RETURN(0, -1, "failed to pwrite `", fn_gzdata);
         }
         return 0;
@@ -373,7 +373,7 @@ private:
         if (gzdata == nullptr) {
             LOG_ERRNO_RETURN(0, -1, "failed to create `", fn_gzdata);
         }
-        if (gzdata->pwrite(gzbuf, gzlen, 0) != gzlen) {
+        if (gzdata->pwrite(gzbuf, gzlen, 0) != (ssize_t)gzlen) {
             LOG_ERRNO_RETURN(0, -1, "failed to pwrite `", fn_gzdata);
         }
         return 0;
@@ -446,7 +446,7 @@ TEST_F(GzCacheTest, cache_store) {
     fread(cbuf1, 1, vsize, fp1);
     fread(cbuf2, 1, vsize, fp2);
     // refill_size is 1MB
-    for (int i = 0; i < vsize; i++) {
+    for (size_t i = 0; i < vsize; i++) {
         if (check_in_interval(i, 0, 1 << 20) ||
             check_in_interval(i, vsize - (1 << 20), vsize) ||
             check_in_interval(i, 5 << 20, 6 << 20)) {
